Refactor code structure and remove redundant code blocks for improved readability and maintainability

+++ resources/update_clone.py
@@ -24,12 +24,14 @@ import shutil
 import xbmc
 import xbmcgui
 import xbmcaddon
+import xbmcvfs
 import re
 import sys
 import time
 import os
 from xml.etree import ElementTree as et
 import fileinput
+import traceback
 
 src_path   = sys.argv[1]
 new_path   = sys.argv[2]
@@ -47,8 +49,7 @@ base_time        = time.time()
 keep_logs        = True if __setting__('logging') == 'true' else False
 
 def lang(id):
-	san = __addon__.getLocalizedString(id).encode( 'utf-8', 'ignore' )
-	return san 
+	return __addon__.getLocalizedString(id)
 
 def log(message, label = '', reset = False):
 	if keep_logs:
@@ -59,67 +60,57 @@ def log(message, label = '', reset = False):
 		start_time   = new_time
 		total_gap    = "%5f" % (new_time - base_time)
 		logmsg       = '%s : %s :: %s ::: %s - %s ' % ('lazyTV addon updater', total_gap, gap_time, label, message)
-		xbmc.log(msg = logmsg)
+		xbmc.log(msg=logmsg, level=xbmc.LOGINFO)
 		base_time    = start_time if reset else base_time
 
 
-
 def errorHandle(exception, trace, new_path=False):
-
 		log('An error occurred while creating the clone.')
 		log(str(exception))
 		log(str(trace))
-
 		dialog.ok('LazyTV', lang(32140),lang(32141))
-		if new_path:
+		if new_path and os.path.exists(new_path):
 			shutil.rmtree(new_path, ignore_errors=True)
 		sys.exit()
 
 
 def Main():
 	try:
-		# remove the existing clone (the settings will be saved in the userdata/addon folder)
-		shutil.rmtree(new_path)
+		if os.path.exists(new_path):
+			shutil.rmtree(new_path)
 
-
-		# copy current addon to new location
 		IGNORE_PATTERNS = ('.pyc','CVS','.git','tmp','.svn')
-		shutil.copytree(src_path,new_path, ignore=shutil.ignore_patterns(*IGNORE_PATTERNS))
+		shutil.copytree(xbmcvfs.translatePath(src_path), new_path, ignore=shutil.ignore_patterns(*IGNORE_PATTERNS))
 
-		# remove the unneeded files
 		addon_file = os.path.join(new_path,'addon.xml')
+		if os.path.exists(os.path.join(new_path,'service.py')): os.remove(os.path.join(new_path,'service.py'))
+		if os.path.exists(addon_file): os.remove(addon_file)
+		if os.path.exists(os.path.join(new_path,'resources','settings.xml')): os.remove(os.path.join(new_path,'resources','settings.xml'))
+		if os.path.exists(os.path.join(new_path,'resources','clone.py')): os.remove(os.path.join(new_path,'resources','clone.py'))
 
-		os.remove(os.path.join(new_path,'service.py'))
-		os.remove(addon_file)
-		os.remove(os.path.join(new_path,'resources','settings.xml'))
-		os.remove(os.path.join(new_path,'resources','clone.py'))
-
-		# replace the settings file and addon file with the truncated one
 		shutil.move( os.path.join(new_path,'resources','addon_clone.xml') , addon_file )
 		shutil.move( os.path.join(new_path,'resources','settings_clone.xml') , os.path.join(new_path,'resources','settings.xml') )
 
 	except Exception as e:
 		ex_type, ex, tb = sys.exc_info()
-		errorHandle(e, tb, new_path)
+		errorHandle(e, traceback.format_tb(tb), new_path)
 
-	# edit the addon.xml to point to the right folder
 	tree = et.parse(addon_file)
 	root = tree.getroot()
 	root.set('id', san_name)
 	root.set('name', clone_name)
-	tree.find('.//summary').text = clone_name
-	tree.write(addon_file)
+	summary = tree.find('.//summary')
+	if summary is not None:
+		summary.text = clone_name
+	tree.write(addon_file, encoding='utf-8', xml_declaration=True)
 
-
-
-	# replace the id on these files, avoids Access Violation
 	py_files = [os.path.join(new_path,'resources','selector.py') , os.path.join(new_path,'resources','playlists.py'),os.path.join(new_path,'resources','update_clone.py')]
-
 	for py in py_files:
-		for line in fileinput.input(py, inplace = 1): # Does a list of files, and writes redirects STDOUT to the file in question
-			print line.replace('script.lazytv',san_name),
+		if not os.path.exists(py): continue
+		with fileinput.FileInput(py, inplace=True) as file:
+			for line in file:
+				print(line.replace('script.lazytv', san_name), end='')
 
-	# stop and start the addon to have it show in the Video Addons window
 	try:
 		log('trying to disable then enable addon')
 		xbmc.executeJSONRPC('{"jsonrpc":"2.0","method":"Addons.SetAddonEnabled","id":1,"params":{"addonid":"%s","enabled":false}}' % san_name)
@@ -132,9 +123,6 @@ def Main():
 	dialog.ok('LazyTV', lang(32149),lang(32147))
 
 if __name__ == "__main__":
-
 	log('Updating started')
-
 	Main()
-
 	log('Updating complete')
